aggiunto ajax post request per time report

+++ database/migrations/2020_05_25_041755_create_time_entry_table.php
@@ -18,7 +18,7 @@ class CreateTimeEntryTable extends Migration
             $table->bigInteger('id_project')->unsigned();
             $table->bigInteger('id_user')->unsigned();
             $table->string('note', 500)->nullable();;
-            $table->decimal('ore', 2, 2);
+            $table->float('ore');
             $table->date('date');
             $table->timestamps();
         });
